+++ src/index.html
@@ -94,7 +94,6 @@ This isn't entirely necessary, but it's considered best practice. Some OpenGL ES
 		<!-- <link id="vtfont" href="https://fonts.googleapis.com/css2?family=VT323&display=swap" rel="stylesheet"> -->
         <link rel="preload" href="https://willard.fun/fonts/VT323.woff2" as="font" type="font/woff2" crossorigin>
         <link rel="shortcut icon" type="image/ico" href="https://willard.fun/assets/minekhan.ico">
-        <link rel="stylesheet" href="index.css">
 	</head>
 	<body>
 	<canvas id="overlay" tabindex="0" width="600" height="600" style="position: absolute; top: 0px; left: 0px"></canvas>
@@ -117,7 +116,6 @@ This isn't entirely necessary, but it's considered best practice. Some OpenGL ES
 		// Place save code here to load your world. Make extra sure you got it copied so you don't paste in the wrong thing and delete your world on accident lol
 		window.loadString = ""
     </script>
-    <script type="module" src="index-test.js"></script>
 	</body>
 </html>
