+++ src/main.js
@@ -8,6 +8,13 @@ import fragmentShaderSrc2D from './shaders/2dFragmentShader.glsl'
 import vertexShaderSrcEntity from './shaders/entityVertexShader.glsl'
 import fragmentShaderSrcEntity from './shaders/entityFragmentShader.glsl'
 
+// imports
+import { seedHash, hash, random, randomSeed, openSimplexNoise, noiseProfile } from "./random.js";
+import { PVector, Matrix, Plane, cross, rotX, rotY, trans, transpose, copyArr } from "./3Dutils.js";
+import { timeString, roundBits } from "./utils.js";
+import { texturesFunc, blockData, BLOCK_COUNT } from "./blockData.js";
+
+window.blockData = blockData
 window.canvas = document.getElementById("overlay")
 window.ctx = window.canvas.getContext("2d")
 window.savebox = document.getElementById("savebox")
@@ -47,923 +54,10 @@ async function MineKhan() {
 		return hash;
 	}
 
-	let setPixel, getPixels;
-
-	const textures = {
-		grassTop: n => {
-			for (let x = 0; x < 16; ++x) {
-				for (let y = 0; y < 16; ++y) {
-					const d = Math.random() * 0.25 + 0.65;
-
-					const r = 0x54 * d;
-					const g = 0xa0 * d;
-					const b = 0x48 * d;
-
-					setPixel(n, x, y, r, g, b);
-				}
-			}
-		},
-		grassSide: n => {
-			const pix = getPixels("0g0g70ordrzz0u30g730wa4vzz0xnyl8f11lrk7315qj7jz1fh47pb6553365533033636350335403653650063306333633300635163503655353653535605335031350330553500033033366333433663663535336655335055335553353530355333033503300333336635353663650660554353355635155305303053556333333366353323553060365553063030663533555365534355335530");
-
-			for (let i = 0; i < pix.length; i += 4) {
-				setPixel(n, i >> 2 & 15, i >> 6, pix[i], pix[i + 1], pix[i + 2], pix[i + 3]);
-			}
-
-			const { random } = Math;
-
-			for (let x = 0; x < 16; ++x) {
-				const m = random() * 4 + 1;
-				for (let y = 0; y < m; ++y) {
-					const d = random() * 0.25 + 0.65;
-					const r = 0x54 * d;
-					const g = 0xa0 * d;
-					const b = 0x48 * d;
-					setPixel(n, x, y, r, g, b);
-				}
-			}
-		},
-		leaves: n => {
-			const { floor, random } = Math;
-
-			for (let x = 0; x < 16; ++x) {
-				for (let y = 0; y < 16; ++y) {
-					const r = 0;
-					const g = floor(random() * 30 + 100);
-					const b = floor(random() * 30);
-					const a = random() < 0.35
-						? 0x0
-						: 0xff;
-
-					setPixel(n, x, y, r, g, b, a);
-				}
-			}
-		},
-		oakPlanks: n => {
-			for (let y = 0; y < 16; ++y) {
-
-				const a = (y & 3) === 3
-					? 0.7
-					: 1.0;
-
-				for (let x = 0; x < 16; ++x) {
-					// these conditions are weird; can some comments be added here?
-					const mid = x === 8 && (y & 7) > 3 && a === 1
-						? 0.85
-						: 1;
-
-					const rit = x === 15 && (y & 7) < 3 && a === 1
-						? 0.85
-						: 1;
-
-					const r = (Math.random() * 0.1 + 0.9) * a * mid * rit;
-
-					setPixel(n, x, y, 190 * r, 154 * r, 96 * r);
-				}
-			}
-		},
-		hitbox: n => {
-			for (let x = 0; x < 16; ++x) {
-				for (let y = 0; y < 16; ++y) {
-					setPixel(n, x, y, 0, 0, 0, 0xff)
-				}
-			}
-		},
-		nothing: n => {
-			for (let x = 0; x < 16; ++x) {
-				for (let y = 0; y < 16; ++y) {
-					setPixel(n, x, y, 0, 0, 0, 0)
-				}
-			}
-		},
-		dirt: "0g0g70ordrzz0u30g730wa4vzz0xnyl8f11lrk7315qj7jz1fh47pb6553365533033636350335403653650063306333633300635163503655353653535605335031350330553500033033366333433663663535336655335055335553353530355333033503300333336635353663650660554353355635155305303053556333333366353323553060365553063030663533555365534355335530",
-		stone: "0g0g40sywflr0wb8hdr0zdjj0f13tzldr3333211210112222221212222220012121001110111222222233232233222111122111212333312223222222211010131223331331112222110010112211122233323223332222212232223332233332021211001212211122222332222233232111232112200101332112211122111321122222222233332222221123322122",
-		logSide: "0g0g60fl1ssf0l5j1fj0qftm2n0wa7mdb14cs7wf16az8xr3143304330341432315230523034133230223052313113324032313232301342413230325232314343134131524131432313422343433143230343243342324331053324324242433315332332414303333432303231430331343230533053135133424052303323531343314231333343143340313114334314134131331432",
-		logTop: "0g0g90l5j1fj0qftm2n0wa7mdb0z2esxr15quebj189da7z1cpma671f7ppfj1hzyayn1012101120110111077776768667777017334454555544811738877776777471164766666666756107576445544674601646636666466471165764655656756116576465464674610657646666566460164764434556756116576666666674610757767777787460175454444444447117776676686677711011101120110211",
-		bedrock: "0g0g509gy58f0e7f7r30o8fd330rkrev31627mkf3111124324211212133434341443012110110111412224232433202422111112014111121134433112221221102211014432344323443410222122211011213234421122344344442110121213211143334134410144431102221123442334402111321134111112343420211101234433211211234421121011044312301123",
-		glass: "0g0g50ybfh8f0znkiyo12rzshr1au95hb1lytipr4444444444444443411111111111111341114111111111104114111111111110413111111111111041111111111111124111111111111110411111111111111241111111111111124111111111111112411111111111111241111111111111123111111111111412411111111111411231111111111111133323222222222233",
-		cobblestone: "0g0g60muaccf0r0pekf0un11q711vr5rz1a8mosf1ef1r0f2144011454313543145330554330132314342143342101321132113232134310021354131154443152453321033543234313433211243215332233213541231321542213533311021543232233121341323231101221044532211235411035532354215434213323154331432332113244323212332143111311232121135432",
-		mossyCobble: "0g0gb0muaccf0mupnnj0p38xdr0r0pekf0rbmj9b0un11q70w1wkxr0y07svz11vr5rz1a8mosf1ef1r0f4199211276438a9619a8812764813858398951644251118533852182851464110531183233866642a176895308948428981498852118851988111851664158385117641484642305126244558824124185442111155214698112124761318998127651764653885847488164588511858685851588531841183352111338a984",
-		stoneBricks: "0g0g70p2gyyn0rkrev30tj2nlr0xf9ou70zdjj0f12psrnj17g8flr5666666666665550645455555444445065454444445534406434554345434330632334544334324053244333345324302222222222222220110011000111111166666650566666663555544065455554544544306435354445544320634344454444343053345433332322305443344322222220222222221111110000111111",
-		mossyStoneBricks: "0g0gc0mupnnj0p2gyyn0p38xdr0rbmj9b0rkrev30tj2nlr0w1wkxr0xf9ou70y07svz0zdjj0f12psrnj17g8flrab3668863b88a680b9a28683a9999332ba926363996a2931b9792a679a979721b737793993697591a72397773632297075225752332277924410441110000444886bbba1abbbb6883aaaa991ba9aaaa6a99a9971b97a7a399aa99351b797992a99997371a689a97777573371a897799755223791722757754000041110004400",
-		bricks: "0g0g90vz62nz0yhavi712oqn7j13rinsv173m8lb193f4zj1b1w1rz1d7u7sv1j1u51b7742888777458777443513444435144410060033100503112256522225565222887747458777474214444315133333151341110600133105522225565522225677458887474588774435344444153444110511331106133325565222225652224777474287774745144444353444441501111106011111065222225655222566",
-		coalOre: "0g0g70ehg7wf0hjr9j30j7xaf30sywflr0wb8hdr0zdjj0f13tzldr6556544543445666554536666553335454331453344511556665655105655644455443346545645556553215542143464552111065105555433556663334466661156543215455645565354100056665354514334663354455555666634106655443366412111054665105556005455554456665566544455555554655555566",
-		ironOre: "0g0g80sywflr0wb8hdr0zdjj0f13tzldr1cpl2bj1gbvabj1o4exa71qwyvb33223211210112333221203333220002121005120011265223332322642322311122110013212312223220762217510131227655432542222100223330001133336523210762122312232021644423332021251001330021122222333301643322110033167666421332542223442122221123332233211122222221322222233",
-		goldOre: "0g0g80sywflr0wb8hdr0zdjj0f13tzldr1x01czj1y6gem71z13ncv1z141z33223211210112333221203333220002121005120011265223332322642322311122110013212312223220762217610131227655432542222100223330001133336523210762122312232021644423332021251001330021122222333301643322110033167655421332642223442122221123332233211122222221322222233",
-		diamondOre: "0g0g80h634zj0sagdtr0sywflr0wb8hdr0zdjj0f13tzldr1845xbz1ndl24f5445433432334555443425555442224343221342233461445554544604544533344332235434534445442764437632353447611054104444322445552223355556145432764344534454243600045554243413223552243344444555523605544332255367611043554604445004344443345554455433344444443544444455",
-		redstoneOre: "0g0g90sywflr0wb8hdr0zdjj0f13oi67z13tzldr15wexa71b68mbj1f24cfz1yr4gsf4224211210112444221204444220002121005120011285224442422832422411122110014212412224220862218610141227655342532222100224440001144448524210862122412242021633324442021251001440021122222444401834422110044168655321442832224332122221124442244211122222221422222244",
-		lapisOre: "0g0ga04hvenz04hvl6n04ihywv066fd3306r2ozj08z4sfz0sywflr0wb8hdr0zdjj0f13tzldr9889877876778999886669999886668787454386777813889889926329989977788776679867978889866428862576797861242398238888723679978767799993189872643386678998687222236258686627661237725788300799668893588779906612366339998700381039799887783339899877788888899888888899",
-		emeraldOre: "0g0g7004swsf06mdmv30sywflr0wb8hdr0zdjj0f13tzldr1ohjdhb5445432232334555443445615442334343223310333422445555225555546133344361324555104445441061243255353445551054434444332232552323355555545461442244534444441053615224243433223310361344444556155551044223455103322553261334455444344441045554455433344554443544444455",
-		coalBlock: "0g0g501e50xr03md24f05ul3b308mtq0v0bf3ri73322122002210012222121000210123321000122000022221001243222202210001233222100210020222221000001220132211001122222022210122343221002110123322210000012123221103200212122210002211232102112210012230002113432123322000123420023221000123210012221001222212212221000",
-		ironBlock: "0g0gb1dawbnj1fj5rlr1hrdssf1m7r1mn1nlyvwf1pa4wsf1qe8xdr1s2ey9r1t6iyv31tqkz5r1ver01r32233333333222232aaaaa9998777772277777777777777105555555666444402aaaaaa999777771277777777777777105555566664444402aa9999977777771277777777777777105555555566644402aaaaaa999977771277777777777777105555556666444402aaaa9999777777127777777777777712222222112111111",
-		goldBlock: "0g0g91kr8um71mphb0f1w77ain1xakkqn1ypvwu71yr43jz1yzk7pb1z0cef31z10mwv2222332223333221285577888776688125664877623324812564877462224471374777462264467137777462267762302877444667762330287444664462232037444664466222613744664466662461364667766666742032667762262776203267762332446261334462332666224132322332662264701101100011001100",
-		diamondBlock: "0g0g90434min061d2in0h634zj0l2fpxb0sagdtr0vckf0f1845xbz1ndl24f1z141z33333223332222331378866777664477138445766432235713845766543335561265666543345546126666543346643203766555446643220376555445543323026555445544333412655445544443541245446644444653023446643343664302346643223554341225543223444335123233223443345601101100011001100",
-		redstoneBlock: "0g0g50vx660v153407319j36671gh49a71runlz34444444444444444433433222334333443342202222232344332311001120234443211012011244443210000000042344302100000011134421100000000122442231000000111344324110210111234432110004001224443321101001110344322211111122224443124211022323443334433223333344444444444444444",
-		lapisBlock: "0g0gd05lqqkf06zt0xr07js8hr07tw35r084kzr308e6ein08e99fj08yjpq708ys8ov0an2j270c0w4cf0dp94hr0fdf5kv98999989cb848484878b78736733677197b8867767623765967877377863366597687666378a63759a77776636a7736197a767676236763596763666a6233625c636636376632775c633687363332371963237873676327186672236338763619763722332677630877633332336667143766633633367708111111610220000",
-		emeraldBlock: "0g0g606lfrb306mdmv307ei5fj07xmdbz0iaro5b10c5ptr0000000000000002055555454551144305000000200002430404555411114243050500000002425305050555445212130405054411421203040405451142120305040411020202130404041122120213042105410112420301212222222242030405510000441213012222222222224301011001110014532333333333333332",
-		tntTop: "0g0g704qh5a70nyecxr13jyl8f14a4flr1d4tukf1otd8u71szz8jj5665566556655665622462246224622462136210012300234431403344004433566006655660566562246204121402246211120000231023443341000111403356651000000506656211020001140224621360200120112344331031140344335661560500611665621462146124612462236223622362234433443344334433",
-		tntSides: "0g0ga07ipw5b0fbhzwf14a4flr1d4tukf1gwtvcv1ldqupr1otd8u71pk09vj1szz8jj1z141z38863886388638863663266326632663266326632663266326632663266326632663266326632663295775577755777799711071550711059745055115157177795717717115505479750571550550759444444444444444433223322332233226632663266326632663266326632663266326632663266323332333233323332",
-		tntBottom: "0g0g413jyl8f14a4flr1d4tukf1otd8u72332233223322332300130013001300130013001300130011111111111111111233223322332233230013001300130013001300130013001111111111111111123322332233223323001300130013001300130013001300111111111111111112332233223322332300130013001300130013001300130011111111111111111",
-		acaciaLogSide: "0g0g60ma181r0oi99fj0pcavi70t8nim70y9464f139ktmn3143304330341432315230523034133230223052313113324032313232301342413230325232314343134131524131432313422343433143230343243342324331053324324242433315332332414303333432303231430331343230533053135133424052303323531343314231333343143340313114334314134131331432",
-		acaciaLogTop: "0g0gb0j7rlz30kvxmgv0ma2nen0o889hb0t8nim713rcxdr16jehof18hk3r31c3oo3j1fprugv1hy2osf432332211231433449999898a889999329556676777766a2395aa999989996910869888888889782197986677668968318688588886886932879868778789784487986876868968448798688887886833869866567789782287988888888968219799899999a96811976766666666691099988988a8899933412233444322114",
-		acaciaPlanks: "0g0g711t8qgv13rcxdr16jehof18hk3r31c3oo3j1fprugv1hy2osf5456666656666652554423455544553345554543333445522112210011010010563666636665443545543432344555543345444255544333001121000011121056654466656666535455554333434332654433334444335300122100012110006556655366646566445533433555544443345432554333330110012221010000",
-		birchLogSide: "0g0g80f1fcov0qqliwv1gxap6n1o60u7z1ptrf271uar6db1uum5mn1z141z36643366634663366346666777634443677744367666636777732100136777737366321101377631211336344363331001266344677766532343777777334556376336775577775777764455577766336336777766634477636777767777577634436633667553210026677763443100000133377761132116331677764336336",
-		birchLogTop: "0g0ga0mk6h3316m5am719xxgqn1cg9ce71f8hx4v1jowirj1nv4jcv1nvimm71tgjy7z1z141z39818811001809889966665657556666816223343444433718627766665666360153655555555645106465334433563580535525555355368154653544545645995465354353563599546535555455358853653323445645115465555555563510646656666676350064343333333336016665565575566688901188999811009",
-		birchPlanks: "0g0g717znmrj19xxgqn1cg9ce71f8hx4v1jowirj1nv4jcv1nvimm75456666656666652554423455544553345554543333445522112210011010010563666636665443545543432344555543345444255544333001121000011121056654466666666535455554333434332654433334444335300122100012110006356655366636566445533433554544443345432554333330110012221010000",
-		darkOakLogSide: "0g0g60besef30dcwlbz0e6y70f0hj7ev30klcs8v0oho0e73143304330341432315230523034133230223052313113324032313232301342413230325232314343134131524131432313422343433143230343243342324331053324324242433315332332414303333432303231430331343230533053135133424052303323531343314231333343143340313114334314134131331432",
-		darkOakLogTop: "0g0gb08ml79b0auqebj0dcmqdb0e6y70f0f117gf0g4r4730h8wirj0hj7ev30kb0idb0lz2akf0n393wf741441100140744779999898a889999419225565666655a1492aa999989995903859888888889681096985566558958408588288885885941869858668689687786985865858958778698588886885844859855256689681186988888888958109699899999a95800965655555555590399988988a8899944701144777411007",
-		darkOakPlanks: "0g0g70bejy0v0dcmqdb0g4r4730h8wirj0kb0idb0lz2akf0n393wf5456666656666652554423455544553345554543333445522112210011010010563666636665443545543432344555543345444255544333001121000011121056654466666666535455554333434332654433334444335300122100012110006356655366636566445533433554544443345432554333330110012221010000",
-		jungleLogSide: "0g0g90h96cxr0htdywv0m9k4xr0ne277j0orqc5b0rkbldr0sxvim70wunksf0ys83cv4688422222666444444114414441001111137510011441866644668842422224146644411143573444422235311010006646442222444886101000466412222444688411068844441122575311222100755444666664355741100112241201661144866643341111623533422111444644111004661664444664122244442222",
-		jungleLogTop: "0g0g90h96cxr0htdywv0rkbldr0ys83cv160fym7188mku71batekf1f77h1b1h5ei2n3212210011213223377776768667777217444454555544812748877776777470164766666666756107576445544674620646646666466472165764655656756336576465464674633657646666566462264764444556756106576666666674600757767777787460175454444444447117776676686677722310122333211013",
-		junglePlanks: "0g0g70sxkd1b0xdxkov160fym7188mku71batekf1f77h1b1h5ei2n5456666656666652554423455544553345554543333445522112210011010010563666636665443545543432344555543345444255544333001121000011121056654466666666535455554333434332654433334444335300122100012110006356655366636566445533433554544443345432554333330110012221010000",
-		spruceLogSide: "0g0g60csc9vj0cskpof0dmmb5r0geuxof0lf4i670nnb4sf3243304330342431325130513034233130113051323223314031323131302341423130315131324343234232514232431323411343433243130343143341314332053314314141433325331331424303333431303132430332343130533053235233414051303313532343324132333343243340323224334324234232332431",
-		spruceLogTop: "0g0g80ix87pb0nnb4sf0p1n6db0qzu7zz0v5xypr0xy569r106bshr11ueyv31012101120110111066665657556666016223343444433711627766665666361153655555555645106465334433563501535525555355361154653533545645115465353353563510546535555455350153653323445645115465555555563510646656666676350164343333333336116665565575566611011101120110211",
-		sprucePlanks: "0g0g80nnb4sf0p1n6db0qzu7zz0v5xypr0xe36rj0xy569r106bshr11ueyv36567777767777762665523566655663356665643333556622112210011010010673777737776553656653532355666653356555266655333001121000011121067765577777777636566665333535332765533335555336300122100012110007367766377737677556633533665655553356532665333330110012221010000",
-		sand: "0g0g61m6x62n1nb9nnj1opn5dr1r80f7j1scbi0v1u0izgf4223213232132313122121130142502432011422222121122331213133132122125213232322122321223332123122121421151211022121212212212111242112322310131232124212221120212231202321232232012311223212331112121213132145321123323230232323221223235332323203223232332321223232",
-		gravel: "0g0g80rufq4f0vqwlbz0zxiprz125i9rz15rvcan1627mkf1d0twqn1dackxr0341152512122521522122312102333522103133522351352232321522512322132741122210253202140133526552213104226515530122553253522311225353521223310256122311652152322132123553102521325022533562113225212132222537415525331025232422215235323311243310351274122232321212",
-		blackConcrete: "0g0g40149on3028826702882yn028dp8f1330112022012232303130022112212111032203010022012122012222321221011223213030101033110011212233120230013131003200032022012002002112233122202312230200102211312102222122132011021201223320211021220121122122321331201102120210001220112022023302312210123220102110",
-		blackWool: "0g0gf0149on3028826702882yn028dp8f03cc2rj03cc3jz03chptr04gg3cv04glqf305kk3y705kpr0f06oo4jj07ss54v08ww5q70a106bj46348b45ab13993299eebb7742570367bc77dc97bcbbdeb996bc768c47cc96ec9749bb115953bb31314c414641672547eebb8879bea7eeb966eebacc45bca9ccbb35ca4657528733577b117949ee7beeccbbecbacb799839349966bb445911796611cd318b9bee92badecb9717bc77cedc97c99beb559b6424c946cd76ac44bc",
-		blueConcrete: "0g0g30c98pof0c98qgv0c9ecqn1000001101000201011001000101120101211111010200000001000110100100220110011001121111001100100100010200102101101000000011110001101010101011020111110000000101121111101100100100010110111010210111001101000100000111110100101111000120000010100000110001200110010002",
-		blueWool: "0g0gj0c98pof0c98qgv0c9ecqn0c9edj30ddcr270ddidbz0ddie4f0ehgrnj0ehgsfz0ehmepr0ehmfi70flkt1b0flqfb30flqg3j0flw2db0flw35r0gpugov0gq02yn0gqbd3335237c44ac12992289iidc6632460256de67ge96cebbhic985ce657e46ee85ie9648cc114942cc21214e413541562436iidc7769dia6iic955iicaee34cea9eecb24ea4546427622466d116949ii7biieecciecaed699728249955cb344911695511eg217c9bii92bahiec9617be67eigf86e99cid449b5324e935eh65ae34de",
-		brownConcrete: "0g0g30pl5gqn0qp3u9r0qp9hbz0211211202222221222122211211122121011102222222212211200222112101121111211111222121222122122221201212121222112122222221220222222122221201121212022122122110221211122221211011221212121010121011111212122221121221222212122112112112211010222111121211220222221122",
-		brownWool: "0g0gi0qp3u9r0qp9hbz0rt7uv30rtdh4v0sxbunz0sxbvgf0u1a8zj0u1fv9b0v5e8sf0v5e9kv0v5jvun0w9i9dr0xdgmwv0xdm9z30yhkni70yhq9rz0zloo3j0zluadb35228d45bd12aa219bhhed7721570257ef77gfa7dfccghdb95df758f47ff95hfb749dd115a52dd21214f413541571526hhed887aehb7hhda55hhdbff25dfbbffdc25fb4557518722576e117b4bhh7chhffddhfdbfe7ab82924bb55dc245b116b5511fg218dachhb1cbghfda717cf77fhgg97fabdhe55ac5214fa25fg75bf24ef",
-		cyanConcrete: "0g0g305objsv05obklb05oh6v31112010122111202100121222111122012212211101111202102221210222122211111211221111011021102021212212220211021022022201210110010212210211110202101010201110112020221222112102111202212110222110211221001010211001221212202222122111221102211100222112011211120220211",
-		cyanWool: "0g0ge05obklb05oh6v305omt4v05omtxb05osg7305oy2gv05p3pj305p9bsv05p9clb05peyv305pkl4v05pklxb05pq87305pq8zj242269347912662167dda95521450145ab55cb659b88cd97649b546b35bb64db7536991146419921113b312431451424dda96656ad75dd9644dd97bb249b77bb9814b73445416521454a115737dd58ddbb99db97ba56762623774498234711474411bc116968dd7187cdb965158b55bdcc65b679da44684213b624bc547b23ab",
-		grayConcrete: "0g0g20ehlvr30flk9a70000000000000000000000000000000000000000000000000000000000000000000000000000100100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000100000000000000000001000000010000",
-		grayWool: "0g0gd0ehlvr30flk9a70flpvjz0flpwcf0flvim70gptwxr0gpzjzz0htxxj30hu3jsv0hu3klb0iy1y4f0iy7ke70iy7l6n333359337803663157cca85531350135ab55cb658b88cc87539b545b35bb53ca7535990036329930103b303330350335cc995556ac75cc8633cc87ba339b77ba8813a73435305532355a005737cc58ccab88cb87ba56753533773498333700574300ac205968cc7087cca865058b55bccc55b679ca33684313a634bc537a33ab",
-		greenConcrete: "0g0g20k2mku70k2s73z0001000100111100101010100101101101101100111010010111001001011000100001100011000100101010110101111010000001011001001000100001010110101110010011010000101011110000010110101000000001010000110101011101000111110011101010011000111100001100011000000111101001101100",
-		greenWool: "0g0gh0k2mku70k2s6bj0k2s73z0l6qjun0l6w64f0l71rlr0mb054v0mb5qm70mb5ren0nf9qf30nffcov0ojdq7z0ojjbpb0ojoxz30pnnbi70pnsxrz0rw0y6746447c45ab1499419aggcb8841580168ce88ee98bebbfgba96ce867e48ee96gda849cc125953cc41114e424642681546ggcc7789cga8ggb966ggbaed45ceaaedbb15da4658517843586c21894agg8bggdebbgebaec89a749449966cb445a116a6621de327c9bgga1bafgdb9828be88egee98e9acgc559b6414d946ef86ad44ce",
-		lightBlueConcrete: "0g0g3090ti4f090z4e70a4xipr2120111111121222222121122112211121121211112221221111102221122212112021111111122111112111112121211222221111122121212121112021122112211121211211111121220111212121111121222001122211222221022122111011221111111021211221211211121112212111111122121111211112222111",
-		lightBlueWool: "0g0gq090z4e70a4xipr0a534zj0a58r9b0b974sf0b9crun0b9ie4f0cdgrnj0cdmepr0cds0zj0dhqein0dhw0sf0dhw1kv0em01dr0em5nnj0fq9o8v0fqfain0gudou70gujb3z0hyhon30hynawv0hynbpb0j2lp8f0j2rbi70k6pptr0latpmn7b57fm89ik26ff63fippnkdd739d04benodepogdkojjppkifblodcfo8doofbpoie8fml21af95mm62428o817b81bd397cppmlffefnpidppkgbbppkioo79loiiookj4aoi8cad93fd55adcn22dh8ippejppookkpokiondfif6f68hhbcmj78ai22cicb22op51fmfjppi3jippokfd2ejoeeoppofdofilpn9agjc738og7copebio78no",
-		lightGrayConcrete: "0g0g20yjgg730yjggzj1000110001001010100010000000011101100010001001011010001100101010001000000000000010100000000100010100000000001010100010001000000000100000110000100001100000011101110001001011001001001001011000000000011001001001001010000000010000000010100110000000001000001001",
-		lightGrayWool: "0g0gi0yjgg730yjggzj0znkgsf0znkhkv10rohdr10roi6711vshz311vsirj11vy5tr12zwjcv12zwk5b1440jy7144670f1584kjj1584lbz16c8l4v16c8lxb17gclq745339e55bd1399329bhhfd7732570257ff78hfa7dfcchhdb95df769f57ff95hfb759ed115953ee31215f514551572537hhed9979fhb7hhda55hhdbff35dfbbgfdc25fb5657529733577f117b5bhh8chhfgddhfdbgf79b93935bb56ec355b117b6511fh319e9chhb2cbhhfd9718cf78fhhg97f9bdhf55ac6325fa36fh75bf35ff",
-		limeConcrete: "0g0g30ppd5a70ppirjz0qth5331121111111100111111111110111110111010211100110111211111211101111112111111111111101110111011110211011011110111110102111110100111000110112112110211112112112111212111111111111111011111011100101111111111212111111111112010011011221111111110111120011110111101111",
-		limeWool: "0g0gg0ppirjz0qth5330qtmrcv0rxl4vz0rxqr5r0t1p4ov0t1uqyn0u5t4hr0u5yqrj0v9x4an0va2qkf0we143j0xi53wf0ym3hfj0ym94hr10uh5of35237c44bb1299228affcb6632460257cd67ed96bdbbefba85bd657d46dd85fca748cb104942cc21214d403540562435ffcb7779cfb6ffb955ffbbdc34bdbaecbb24cb4546427622465c01694aff7bffcebbfdbbec69a728249955cb344a105a5501ce207c9bffa2bbefcb9607bd77dfee86d9abfc449b5324c935de75bc34cd",
-		magentaConcrete: "0g0g419jw2rj1anugan1ao02kf1ao03cv2233313332221332333333233213322333223332331233333313323103121223222131233211303233130230322231223302323103002220333103232232223333313231221331233333231232212303302330122232333333332123320111033031332321233232303322232223332223231233212323332233222231230323",
-		magentaWool: "0g0gs1ao02kf1ao03cv1brygvz1bs435r1bs43y71cw2hhb1cw83r31cw84jj1e06i2n1e0c4cf1e0c54v1f4ainz1f4g4xr1f4g5q71g8ej9b1g8k5j31g8k6bj1hcijun1hco64f1hco6wv1igmkfz1igs6pr1igs7i71jkw7b31jkw83j1kp08ov1lt49a71mxdwxr6945dl77ij14ff43ehrrmjba537a039cmoacqogajojjqrjhe9koa9do7booe9rnhc7elk118f74ll41317o7169719a275arrmkddcfmribrrjg99rrjion57koihonjj38ni798a72db448bam11bh7hrrcjrrnojjrojiomafhd4e47hh99lj578h11ah9911nq41dlfjrrh2jiqrnjfa1cjoccorqpebofhkrm78gj9537ng59oqc9in57mo",
-		orangeConcrete: "0g0g21p59iwv1q97wfz1011110111101111111011111110000111101110100111111110011111011001001101110110111111111110111111011111101110111111110111111111111111011110111101111011101101101111010010111111110110101110010011101111111010101001101111111011111110111111111001111111101111111111",
-		orangeWool: "0g0gm1q97wfz1rdbw8v1rdbx1b1rdhjb31shfwu71shfxmn1tljxfj1tljy7z1tlpkhr1upnytb1uptl331uptlvj1vtrzen1vtxlof1vtxmgv1wy1m9r1wy1n271wy79bz1wycwe71wyijgf1wyo5q71wytssf46239f55de02aa219dllhe8731570268hi79kib7eieelled96fi779i58ii96lhd859ff005a52ff20205i504650671537llgf998ahld8lleb66lledih35fiddihee25hd5757519822587h008c5dll9ellhieeliedih7ad92925cc67fe355d007d7600hk209faelld1edllhea709ei89ilkj98iadflh55be7315hb37il86dh35hi",
-		pinkConcrete: "0g0g31ltks8v1mxj5rz1mxj6kf1122212121112211111211211211122112111221111111121211212111122211111112112121112111111111111112111212221121111111212211112101110212211211122111211112212111211110211211211211111111121212221211121111112212211121211111211212211111111111121112111211212122122111",
-		pinkWool: "0g0gv1mxj6kf1o1hk3j1o1n6db1p5r6yn1q9v6rj1q9v7jz1rdz7cv1re4uf31si37y71si8u7z1tm78jj1tmcutb1tmcvlr1uqgven1uqmigv1uqs4qn1vuqj271vuw5bz1vuw64f1vv1se71vv7fgf1vvd1q71vvd2in1vviosf1vvipkv1vvobun1vvocn31vvtywv1vvzlz31vw588v1vw591b6a46em78jk25gg53fjuuokdc638c03adoqcesqhckqkktukjfalqcbeq7dqqfaupjd7fml219g83mm52327q716a71ac386buunleedgoujduukhaauukjqp68lqjjrpkk39pj7b9c83ed439dbo12di7juuekuuprkkuqkjrocgje5f57iiabmk679j22bjba12ps31emgkuuj3kjtupkgc1ekqdequssfdqgjluo89hkb637ph6bqtdajp67oq",
-		purpleConcrete: "0g0g40qo16v30rrzke70rrzl6n0rs57gf3222322033323333333312320220233231302222232332333202332233033330333322320012333323333330233022223332223332223232323303323233332332021220323110033223333322032332312323332322322202232223033232221223233023333332022233332332233232023333233223032323332323323332",
-		purpleWool: "0g0gm0rrzl6n0rs57gf0sw3kzj0sw3lrz0u01zb30u07lkv0u07mdb0v45zwf0w84dfj0w84e7z0xc8e0v0xc8etb0yg6scf0ygcem70zkasxr10o96gv10o979b10oetj311sd72711sd7un12wh8fz154uxhb47349e45ce13aa32acllfe8742570279fh79jhb7ehddkleca7eh779h48hha7lgc94aee116a52ee31214h414741771547llfe999aflc8lleb77llechg45ehcchged26gc4767519832687f118c4cll9dllgheelhechf7ac93a34cc77ed446c117c7711gj219eadllc1dcklgea719dh99hljia8hacelf56bd7424gb47hk97cg44fh",
-		redConcrete: "0g0g112voa9r0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-		redWool: "0g0gd12voa9r13zmnsv153l1bz167jev3167p14v17bnenz18fls7319jprzz1ano5j31ano6bj1brmjun1brs64f1f3yku7241258227712662157cc9754212401459b45bb647b77bc77548b445b25bb54ca7525881036218821112b202420441224cc8855569c75cc7644cc77ba228b77ba7713a724342155113549015627cc57ccab77cb77b946752522664487223711474401ab105867cc7177bca764057b55bcbb55b678c923674212a624bb547a229b",
-		whiteConcrete: "0g0g31ktui9r1ku04jj1lxyi2n1121212211221121111221111121122111211111121112111212112211111211212222111111221112211112101110211111122211211211111111212221111112212121212222111121122211111111222111111112121222121111111111211121221112221111121222121112212122111221211212111121221211221212",
-		whiteWool: "0g0gn1lxyi2n1lxyiv31ly454v1n22inz1n22jgf1n285q71o66j9b1o66k1r1o6c6bj1paajun1qeekfz1ricyrj1riil1b1smgykf1smmlmn1tqkz5r1tqqm7z1uuozr31uuumtb1vyt0cf1vyynen1x2x0xr1y711j37a56dj89hi26ee63egmmliba639a04acllacmlfailiimmigeajlaadl8blleamlgc8ejj219e94jj62429l817a81aa396ammkjddcelmhbmmifaammihll69jlhgllii49lh8a9a93db549bal12bg8gmmcimmlliimlihllaegd6e68ggaaji689g22agaa12lm41djeimmg3ihmmliea1cilcclmmmeblegjml99fia638lf6almcahl68ll",
-		yellowConcrete: "0g0g41to1w5b1us09of1us5vy71us5wqn2111111122222221232021021213122222211020112110121211022201012212122222312211221122213110202011133121132211112221221211211111112120231210211111201132001211212111211102220222021102211223011100222121202222222132111211111112111112021121120222221222011012122121",
-		yellowWool: "0g0gj1us5vy71usbj0f1vw9wjj1vwfitb1vwfjlr1vwl5vj1vwl6nz1vwqsxr1x0p6gv1x0utj31x10fsv1x10glb1x163nj1y54h6n1y5a48v1y5a51b1y5frb31y5fs3j1y5lf5r45239e44bc0299219biiec7731470158eg78hga7cgcchicb95dg759g47gg95ifb849ed005942ee20104g404540571436iied9989eib7iica55iicbgf34dgbbgfcc15fb4557419722576e007b4bii8ciifgccigcbge79b92924bb55ec345b006b5500fh209e9ciib1cbhifc9708cg88gihh97g9bdie45ac5314fa35gh85bf34eg",
-		bookshelf: "0g0gt03fxnnj04laqdb0a0ot1b0b6j6db0c8r6db0deww730df88ov0egz6rj0gpo9330ht5kov0k20av30nf40zj0pnc1dr0qoh8fz0sy416n0w8kcn30wc5n9b0yicu0v11vb08v11vskcf13z03jz16atkvz17fehvj1as1ce71czhmv31e50qv31g6nvgf1gbtpfj1ks44qnommllhlllmmmmlmoo44477444772534oo799kk999ni9637olkffqk99gpjk631lsqffqk64gnik631soqfdqf6gcngf651ookd9kf3c7igf350mlorrsssroorrroolollmmhmllmmmllhoo77227cb7427724ooqk763gck99cb97olpj063gckq4gc10hsqk063a8fqkgcpjsoqf065a87kqgc10omkfe35ccefkcb03oorrsssroorrrrooo",
-		netherBricks: "0g0g706o77cv08w9lhb0b4bzlr0dcedq70fkb4sf0gof6670iwn6kf0000000000000000556515666555165532340433333405333432133223321333000111000001110015666544155545540543333306433333143333221433243211000001110000014416641644164414330433063304330423143314321333140111011101110111464414666644166433330433344303333322033333330333",
-		redNetherBricks: "0g0g70c7i51b0efew3j0fjiwov0ive1a70nbiubj0svgd1b0v3ir5r0000000000000000556515666555165532340433333405333432133223321333000111000001110015666544155545540543333306433333143333221433243211000001110000014416641644164414330433063304330423143314321333140111011101110111464414666644166433330433344303333322033333330333",
-		netherQuartzOre: "0g0gc0hs27zz0m7vq4f0m81d6n0ncb0u70rsft330u2845b0v4m7sv10p0npb1bug2671e4e1vj1n1fv9b1smb9xb1312426644044643344320246610442164643068a921424676466baa866314246766ba8674662643166895678a8544641466565ba95146866310289856642b9742469b814466ab862467ba624646b952364a950268665602426852108b61623214642649a9164941364667689238976124667423804646442246624630246644",
-		netherrack: "0g0g70hs27zz0m7vq4f0m81d6n0ncb0u70rsft330v4m7sv10p0npb1312425544044543344320245210442154543012432142456545423545531424565424566455254315431566554244541452054554214565531024541154246642454221445541452456540245455112354554025655200242254210254102321454254122154541354556512234565124556423204545442245524530245544",
-		netherWartBlock: "0g0g60of09vj0suze9r0xana4f1539r7j1brbain1g7rcvz0122001222532010222235202110022112111222202202320122201003202242042220021211122103224300222201201022220122252002210210211223102220000221222200222022221132210202012201104100222102201012211021042230122022012022212012022002220222202202220234022205220322022302",
-		quartzBlockBottom: "0g0g31sm5mv31tq9ngf1tqfbb32222221111000222121111000002222111100000012222110000001122221110001111000110000011100000110000011111222200001111222222211111111222211000112222220110012222222211222222222210000022222211000000222222110110112222111100112222221100101122222111101222222222210000",
-		quartzBlockSide: "0g0g61p9z6kf1qe37y71sm5mv31tq9ngf1tqfbb31uujcov5555555555555555543333222224444153322222234444304222223344433330423333222332222153322222332222215333444422223330544444433333333054433222334444405332234444444430544444444432222154444433222222405444332332334440533322334444443042323344444333311100011111110000",
-		quartzBlockTop: "0g0g61p9z6kf1qe37y71sm5mv31tq9ngf1tqfbb31uujcov5555555555555555543333222224444153322222234444304222223344433330423333222332222153322222332222215333444422223330544444433333333054433222334444405332234444444430544444444432222154444433222222405444332332334440533322334444443042323344444333311100011111110000",
-		quartzPillar: "0g0g41qe37y71sm5mv31tqfbb31uujcov0202031312130212020302130313031202131313120313020313130213031313131303120302121313020312031203131302131313020303120313031303130312130303131303120213131302131212131312130202120313021213120302031302120313031303130202130303131303120313031213120212131302121202",
-		quartzPillarTop: "0g0g51p9z6kf1qe37y71sm5mv31tqfbb31uujcov1021223242413020222122211221122002223434434322223123443344443211113412222221431242442344443244222243242222424423423324244142441222442424324234143244242211423322224423334432442421341222222143111123444444443213222234444443222002211222122112221203142423220200",
-		chiseledQuartzBlock: "0g0g61p9z6kf1qe37y71sm5mv31tq9ngf1tqfbb31uujcov0444404333022440043331222204444003322111000444300222223004443331023011101110022103315544432202200110441100220000554444144033344544433204314444440111230011441101054044444432154104400110111004401444454005544440133321100004443002323155440333300444414444132220",
-		chiseledQuartzBlockTop: "0g0g51p9z6kf1qe37y71tq9ngf1tqfbb31uujcov0334314232133420021100022000112002044203204421200204120420412021020113032020003102233304212233200000000421110000323433442223344422222222222222220000001321000001044343142143222104011204214110201302130420313020120234032023403002111003201100400222314222133420",
-		chiseledStoneBricks: "0g0g70oigd8f0qqoef30sz21vj0xf6vpb0yjgikf11vskcf17g711b6666655666566562644424444224444054110000111123205206665566561530640654434442154064064322334315405205430005421630540632000643062052054300053206306215435655420630641644545432053064163232322215406421110000011530556655666655543053343234433232202000000000000000",
-		smoothStone: "0g0g70wb8hdr0yjgikf12zwkxr17gcnb318kgnwf1asop331d0wq9r1011002200220121054556666553445204334454444555501666455566655442255444545666645006555555554343412556664665545550243343443434466116656545555455620555554566656661054544334454454215555666665566622444566454433452165445556555455104456665566544401022112221000121",
-		soulSand: "0g0g60egz94v0ht5n270k181z30oho3jz0sxyiv30xeekfz3130033202100434433135531100234004241451321012411423432542212230032442154212211332234104322433212002314322315332220023322345513032322211233445304302332013540423202342330242132302341043213032020354044321232020431445421112202350345131123302354202432103212415",
-		glowstone: "0g0g80u1fugv0v5phbz10pn01r11uj6db1kqv1tr1x1ncov1y6jzlr1y711j35420045440276423654011240346654024032201131054033034542017510033014766405664136502566552454236751004554020132450531024010352320465230213276501576442654346640335420454220452013000320356304012311354016754036520357660354215764034654201003242013154216521100013",
-		andesite: "0g0g60sywflr0wb8hdr0yjgikf11vsl4v17gcnb31asolxb2114332104423221245323445330122411334410122443224410133445433231235442223333211433323243210013432324542104543333423310154332333433114143343224533143342221154333332321101544332201331254422332333443443233212331543333334533112233301244323223541321322323234432",
-		diorite: "0g0g60xfchz311vskcf19of1fj1gd8su71ktouf31smgykf1222352233524355201322135345543151355101325543235255312553213225351243153532255245355541332335235332455214554102243123234555521232203255443244250155213551234541244310134335513055510144221512352542543554223223212343125454353242115521455321245235554124421255",
-		granite: "0g0ga0l56fi70plgu0v0yhw0sf1424tmn17eb8cf1aqna4f1e3al8f1e3x7un1jnje2n1qczoxr7434244344046445431548244224422341444335249241464143446145532344234232113754334244255342143347242445524442555322424474302155243346344255522344245535242552464245557314243844212555415442024425030323244424324439435524312355344448553215435547245544733324244425",
-		polishedAndesite: "0g0g90oilzi70sz23gf0v7a3270yjgjcv10ru60v11vyakf11w3ugv19okmwv1e4v30f7878888888788785843334442443134084444443376644307416763343446440844442456433334083344333424433318444634666554320833433332123444184562444466776608434433443344540844667633366444081444444664466208466654333134760844433334465334083333664444334406100000000110000",
-		polishedDiorite: "0g0g80qqu0ov0xfi5tr11vy6m71asubcv1f94t1b1jpkttr1o60w731ri7bpb7776677777666774764556657643564166764466577434506565333565454441675334576435457176544567733456607566767655346450744765756564534063556655467534516566544565544761664334655465666075673555775644517557654375466340745654576435744072445566545654512101110001100110",
-		polishedGranite: "0g0g90l56fi70plgu0v0yhw0sf1424tmn17eb8cf1aqna4f1e3al8f1e3x7un1jnje2n8888786866866565843344345444334374744554442446418444243334553342844444453344443262343344443734428345543644444431844444423445544263422344554433636444433443344442734444444464334284473344244442318455444445543341633443264434455264443444334447412111001110111100",
-	}
-
-	const blockData = [
-		{
-			name: "air",
-			id: 0,
-			textures: [],
-			transparent: true,
-			shadow: false,
-			solid: false
-		},
-		{
-			name: "grass",
-			textures: ["dirt", "grassTop", "grassSide"],
-		},
-		{ name: "dirt" },
-		{ name: "stone" },
-		{ name: "bedrock" },
-		{ name: "sand" },
-		{ name: "gravel" },
-		{
-			name: "leaves",
-			transparent: true,
-		},
-		{
-			name: "glass",
-			transparent: true,
-			shadow: false,
-		},
-		{ name: "cobblestone" },
-		{ name: "mossyCobble" },
-		{ name: "stoneBricks" },
-		{ name: "mossyStoneBricks" },
-		{ name: "bricks" },
-		{ name: "coalOre" },
-		{ name: "ironOre" },
-		{ name: "goldOre" },
-		{ name: "diamondOre" },
-		{ name: "redstoneOre" },
-		{ name: "lapisOre" },
-		{ name: "emeraldOre" },
-		{ name: "coalBlock" },
-		{ name: "ironBlock" },
-		{ name: "goldBlock" },
-		{ name: "diamondBlock" },
-		{ name: "redstoneBlock" },
-		{ name: "lapisBlock" },
-		{ name: "emeraldBlock" },
-		{ name: "oakPlanks" },
-		{
-			name: "oakLog",
-			textures: ["logTop", "logSide"],
-		},
-		{ name: "acaciaPlanks" },
-		{
-			name: "acaciaLog",
-			textures: ["acaciaLogTop", "acaciaLogSide"],
-		},
-		{ name: "birchPlanks" },
-		{
-			name: "birchLog",
-			textures: ["birchLogTop", "birchLogSide"],
-		},
-		{ name: "darkOakPlanks" },
-		{
-			name: "darkOakLog",
-			textures: ["darkOakLogTop", "darkOakLogSide"],
-		},
-		{ name: "junglePlanks" },
-		{
-			name: "jungleLog",
-			textures: ["jungleLogTop", "jungleLogSide"],
-		},
-		{ name: "sprucePlanks" },
-		{
-			name: "spruceLog",
-			textures: ["spruceLogTop", "spruceLogSide"],
-		},
-		{ name: "whiteWool" },
-		{ name: "orangeWool" },
-		{ name: "magentaWool" },
-		{ name: "lightBlueWool" },
-		{ name: "yellowWool" },
-		{ name: "limeWool" },
-		{ name: "pinkWool" },
-		{ name: "grayWool" },
-		{ name: "lightGrayWool" },
-		{ name: "cyanWool" },
-		{ name: "purpleWool" },
-		{ name: "blueWool" },
-		{ name: "brownWool" },
-		{ name: "greenWool" },
-		{ name: "redWool" },
-		{ name: "blackWool" },
-		{ name: "whiteConcrete" },
-		{ name: "orangeConcrete" },
-		{ name: "magentaConcrete" },
-		{ name: "lightBlueConcrete" },
-		{ name: "yellowConcrete" },
-		{ name: "limeConcrete" },
-		{ name: "pinkConcrete" },
-		{ name: "grayConcrete" },
-		{ name: "lightGrayConcrete" },
-		{ name: "cyanConcrete" },
-		{ name: "purpleConcrete" },
-		{ name: "blueConcrete" },
-		{ name: "brownConcrete" },
-		{ name: "greenConcrete" },
-		{ name: "redConcrete" },
-		{ name: "blackConcrete" },
-		{
-			name: "bookshelf",
-			textures: ["oakPlanks", "bookshelf"]
-		},
-		{ name: "netherrack" },
-		{ name: "soulSand" },
-		{
-			name: "glowstone",
-			lightLevel: 15
-		},
-		{ name: "netherWartBlock" },
-		{ name: "netherBricks" },
-		{ name: "redNetherBricks" },
-		{ name: "netherQuartzOre" },
-		{
-			name: "quartzBlock",
-			textures: ["quartzBlockBottom", "quartzBlockTop", "quartzBlockSide"]
-		},
-		{
-			name: "quartzPillar",
-			textures: ["quartzPillarTop", "quartzPillar"]
-		},
-		{
-			name: "chiseledQuartzBlock",
-			textures: ["chiseledQuartzBlock", "chiseledQuartzBlockTop"]
-		},
-		{ name: "chiseledStoneBricks" },
-		{ name: "smoothStone" },
-		{ name: "andesite" },
-		{ name: "polishedAndesite" },
-		{ name: "diorite" },
-		{ name: "polishedDiorite" },
-		{ name: "granite" },
-		{ name: "polishedGranite" },
-		{ name: "light", textures: "nothing", lightLevel: 15, solid: false, icon: "glass", transparent: true, shadow: false }
-		// I swear, if y'all don't stop asking about TNT every 5 minutes!
-		/* {
-			name: "tnt",
-			textures: ["tntBottom", "tntTop", "tntSides"]
-		},*/
-	];
-	window.blockData = blockData
-
-	const BLOCK_COUNT = blockData.length;
-
-	// Set defaults on blockData
-	for (let i = 1; i < BLOCK_COUNT; ++i) {
-		const data = blockData[i];
-		data.id = i;
-
-		if ( !("textures" in data) ) {
-			data.textures = new Array(6).fill(data.name);
-		}
-		else if (typeof data.textures === "string") {
-			data.textures = new Array(6).fill(data.textures);
-		}
-		else {
-			const { textures } = data;
-
-			if (textures.length === 3) {
-				textures[3] = textures[2];
-				textures[4] = textures[2];
-				textures[5] = textures[2];
-			}
-			else if (textures.length === 2) {
-				// Top and bottom are the first texture, sides are the second.
-				textures[2] = textures[1];
-				textures[3] = textures[2];
-				textures[4] = textures[2];
-				textures[5] = textures[2];
-				textures[1] = textures[0];
-			}
-		}
-
-		data.transparent = data.transparent || false;
-
-		data.shadow = data.shadow !== undefined
-			? data.shadow
-			: true;
-
-		data.lightLevel = data.lightLevel || 0;
-		data.solid = data.solid !== undefined ? data.solid : true
-		data.icon = data.icon || false
-	}
-
-	// implementation of xxHash
-	const {
-		seedHash,
-		hash
-	} = (() => {
-		// closure around mutable `seed`; updated via calls to `seedHash`
-
-		let seed = Math.random() * 2100000000 | 0;
-
-		const PRIME32_2 = 1883677709;
-		const PRIME32_3 = 2034071983;
-		const PRIME32_4 = 668265263;
-		const PRIME32_5 = 374761393;
-
-		const seedHash = s => {
-			seed = s | 0;
-		}
-
-		const { imul } = Math;
-
-		const hash = (x, y) => {
-			let h32 = 0;
-
-			h32 = seed + PRIME32_5 | 0;
-			h32 += 8;
-
-			h32 += imul(x, PRIME32_3);
-			h32 = imul(h32 << 17 | h32 >> 32 - 17, PRIME32_4);
-			h32 += imul(y, PRIME32_3);
-			h32 = imul(h32 << 17 | h32 >> 32 - 17, PRIME32_4);
-
-			h32 ^= h32 >> 15;
-			h32 *= PRIME32_2;
-			h32 ^= h32 >> 13;
-			h32 *= PRIME32_3;
-			h32 ^= h32 >> 16;
-
-			return h32 / 2147483647;
-		};
-
-		return {
-			seedHash,
-			hash
-		};
-	})();
-
 	let world;
 	let worldSeed;
-
-	class Marsaglia {
-		// from http://www.math.uni-bielefeld.de/~sillke/ALGORITHMS/random/marsaglia-c
-
-		nextInt() {
-			const { z, w } = this;
-
-			this.z = 36969 * (z & 65535) + (z >>> 16) & 0xFFFFFFFF;
-			this.w = 18000 * (w & 65535) + (w >>> 16) & 0xFFFFFFFF;
-
-			return ((this.z & 0xFFFF) << 16 | this.w & 0xFFFF) & 0xFFFFFFFF;
-		}
-
-		nextDouble() {
-			const i = this.nextInt() / 4294967296;
-
-			const isNegative = i < 0 | 0; // cast to 1 or 0
-
-			return isNegative + i;
-		}
-
-		constructor(i1, i2) { // better param names
-			this.z = i1 | 0 || 362436069;
-			this.w = i2 || hash(521288629, this.z) * 2147483647 | 0;
-		}
-	}
-
-	// The noise and random functions are copied from the processing.js source code; these others are polyfills made by me to avoid needing to remove all the pjs draw calls
-
-	const {
-		randomSeed,
-		random
-	} = (() => {
-		// closure around mut `currentRandom`
-
-		let currentRandom = null;
-
-		const randomSeed = seed => {
-			currentRandom = new Marsaglia(seed);
-		};
-
-		const random = (min, max) => {
-			if (!max) {
-				if (min) {
-					max = min;
-					min = 0;
-				}
-				else {
-					min = 0;
-					max = 1;
-				}
-			}
-
-			return currentRandom.nextDouble() * (max - min) + min;
-		};
-
-		return {
-			randomSeed,
-			random
-		};
-	})();
-
-	class PerlinNoise {
-		// http://www.noisemachine.com/talk1/17b.html
-		// http://mrl.nyu.edu/~perlin/noise/
-
-		static grad3d(i, x, y, z) {
-			const h = i & 15; // convert into 12 gradient directions
-
-			const u = h < 8
-				? x
-				: y;
-
-			const v = h < 4
-				? y
-				: h === 12 || h === 14
-					? x
-					: z;
-
-			return ((h & 1) === 0 ? u : -u) + ((h & 2) === 0 ? v : -v)
-		}
-
-		static grad2d(i, x, y) {
-			const v = (i & 1) === 0
-				? x
-				: y;
-
-			return (i & 2) === 0
-				? -v
-				: v;
-		}
-
-		static grad1d(i, x) {
-			return (i & 1) === 0
-				? -x
-				: x;
-		}
-
-		static lerp(t, a, b) {
-			return a + t * (b - a);
-		}
-
-		// end of statics
-
-		// prototype functions:
-		noise3d(x, y, z) {
-			const { floor } = Math;
-
-			const X = floor(x) & 0xff;
-			const Y = floor(y) & 0xff;
-			const Z = floor(z) & 0xff;
-
-			x -= floor(x);
-			y -= floor(y);
-			z -= floor(z);
-
-			const fx = (3 - 2 * x) * x * x;
-			const fy = (3 - 2 * y) * y * y;
-			const fz = (3 - 2 * z) * z * z;
-
-			const { perm } = this;
-
-			const p0 = perm[X] + Y;
-			const p00 = perm[p0] + Z;
-			const p01 = perm[p0 + 1] + Z;
-			const p1 = perm[X + 1] + Y;
-			const p10 = perm[p1] + Z;
-			const p11 = perm[p1 + 1] + Z;
-
-			const { lerp, grad3d } = PerlinNoise;
-
-			return lerp(
-				fz,
-				lerp(
-					fy,
-					lerp(
-						fx,
-						grad3d(perm[p00], x, y, z),
-						grad3d(perm[p10], x - 1, y, z)
-					),
-					lerp(
-						fx,
-						grad3d(perm[p01], x, y - 1, z),
-						grad3d(perm[p11],x - 1, y - 1, z)
-					)
-				),
-				lerp(
-					fy,
-					lerp(
-						fx,
-						grad3d(perm[p00 + 1], x, y, z - 1),
-						grad3d(perm[p10 + 1], x - 1, y, z - 1)
-					),
-					lerp(
-						fx,
-						grad3d(perm[p01 + 1], x, y - 1, z - 1),
-						grad3d(perm[p11 + 1], x - 1, y - 1, z - 1)
-					)
-				)
-			);
-		}
-
-		noise2d(x, y) {
-			const { floor } = Math;
-
-			const X = floor(x) & 0xff;
-			const Y = floor(y) & 0xff;
-
-			x -= floor(x);
-			y -= floor(y);
-
-			const { perm } = this;
-			const fx = (3 - 2 * x) * x * x;
-			const fy = (3 - 2 * y) * y * y;
-			const p0 = perm[X] + Y;
-			const p1 = perm[X + 1] + Y;
-
-			const { lerp, grad2d } = PerlinNoise;
-
-			return lerp(
-				fy,
-				lerp(
-					fx,
-					grad2d(
-						perm[p0],
-						x,
-						y
-					),
-					grad2d(
-						perm[p1],
-						x - 1,
-						y
-					)
-				),
-				lerp(
-					fx,
-					grad2d(
-						perm[p0 + 1],
-						x,
-						y - 1
-					),
-					grad2d(
-						perm[p1 + 1],
-						x - 1,
-						y - 1
-					)
-				)
-			);
-		}
-
-		constructor(seed) {
-			if (seed === undefined) {
-				throw new TypeError("A value for `seed` parameter was not provided to `PerlinNoise`");
-			}
-
-			const rnd = new Marsaglia(seed);
-
-			// permutation
-			this.perm = new Uint8Array(0x200);
-
-			// generate permutation
-			const { perm } = this;
-
-			// fill 0x0..0x100
-			for (let i = 0; i < 0x100; ++i) {
-				perm[i] = i;
-			}
-
-			for (let i = 0; i < 0x100; ++i) {
-				const j = rnd.nextInt() & 0xFF;
-				const t = perm[j];
-				perm[j] = perm[i];
-				perm[i] = t;
-			}
-
-			// copy to avoid taking mod in perm[0]
-			// copies from first half of array, into the second half
-			perm.copyWithin(0x100, 0x0, 0x100);
-		}
-	}
-
-	const noiseProfile = {
-		generator: undefined,
-		octaves: 4,
-		fallout: 0.5,
-		seed: undefined,
-		noiseSeed(seed) {
-			this.seed = seed;
-			this.generator = new PerlinNoise(noiseProfile.seed);
-		},
-		noise(x, y, z) {
-			const { generator, octaves, fallout } = this;
-
-			let effect = 1,
-				sum = 0;
-
-			for (let i = 0; i < octaves; ++i) {
-				effect *= fallout;
-
-				const k = 1 << i;
-
-				let temp;
-				switch (arguments.length) {
-					case 1: {
-						temp = generator.noise1d(k * x);
-						break;
-					} case 2: {
-						temp = generator.noise2d(k * x, k * y);
-						break;
-					} case 3: {
-						temp = generator.noise3d(k * x, k * y, k * z);
-						break;
-					}
-				}
-
-				sum += effect * (1 + temp) / 2;
-			}
-
-			return sum;
-		}
-	};
-
 	let caveNoise;
-	// Copied and modified from https://github.com/blindman67/SimplexNoiseJS
-	function openSimplexNoise(clientSeed) {
-		const toNums = function(s) {
-			return s.split(",").map(function(s) {
-				return new Uint8Array(s.split("").map(function(v) {
-					return Number(v)
-				}))
-			})
-		}
-		const decode = function(m, r, s) {
-			return new Int8Array(s.split("").map(function(v) {
-				return parseInt(v, r) + m
-			}))
-		}
-		const toNumsB32 = function(s) {
-			return s.split(",").map(function(s) {
-				return parseInt(s, 32)
-			})
-		}
-		const NORM_3D = 1.0 / 206.0
-		const SQUISH_3D = 1 / 3
-		const STRETCH_3D = -1 / 6
-		var base3D = toNums("0000110010101001,2110210120113111,110010101001211021012011")
-		const gradients3D = decode(-11, 23, "0ff7mf7fmmfffmfffm07f70f77mm7ff0ff7m0f77m77f0mf7fm7ff0077707770m77f07f70")
-		var lookupPairs3D = function() {
-			return new Uint16Array(toNumsB32("0,2,1,1,2,2,5,1,6,0,7,0,10,2,12,2,41,1,45,1,50,5,51,5,g6,0,g7,0,h2,4,h6,4,k5,3,k7,3,l0,5,l1,5,l2,4,l5,3,l6,4,l7,3,l8,d,l9,d,la,c,ld,e,le,c,lf,e,m8,k,ma,i,p9,l,pd,n,q8,k,q9,l,15e,j,15f,m,16a,i,16e,j,19d,n,19f,m,1a8,f,1a9,h,1aa,f,1ad,h,1ae,g,1af,g,1ag,b,1ah,a,1ai,b,1al,a,1am,9,1an,9,1bg,b,1bi,b,1eh,a,1el,a,1fg,8,1fh,8,1qm,9,1qn,9,1ri,7,1rm,7,1ul,6,1un,6,1vg,8,1vh,8,1vi,7,1vl,6,1vm,7,1vn,6"))
-		}
-		var p3D = decode(-1, 5, "112011210110211120110121102132212220132122202131222022243214231243124213241324123222113311221213131221123113311112202311112022311112220342223113342223311342223131322023113322023311320223113320223131322203311322203131")
-		const setOf = function(count) {
-			var a = [], i = 0;
-			while (i < count) {
-				a.push(i++)
-			}
-			return a
-		}
-		const doFor = function(count, cb) {
-			var i = 0;
-			while (i < count && cb(i++) !== true);
-		}
 
-		function shuffleSeed(seed,count){
-			seed = seed * 1664525 + 1013904223 | 0
-			count -= 1
-			return count > 0 ? shuffleSeed(seed, count) : seed
-		}
-		const types = {
-			_3D : {
-				base : base3D,
-				squish : SQUISH_3D,
-				dimensions : 3,
-				pD : p3D,
-				lookup : lookupPairs3D,
-			}
-		}
-
-		function createContribution(type, baseSet, index) {
-			var i = 0
-			const multiplier = baseSet[index ++]
-			const c = { next : undefined }
-			while(i < type.dimensions) {
-				const axis = "xyzw"[i]
-				c[axis + "sb"] = baseSet[index + i]
-				c["d" + axis] = - baseSet[index + i++] - multiplier * type.squish
-			}
-			return c
-		}
-
-		function createLookupPairs(lookupArray, contributions){
-			var i
-			const a = lookupArray()
-			const res = new Map()
-			for (i = 0; i < a.length; i += 2) {
-				res.set(a[i], contributions[a[i + 1]]);
-			}
-			return res
-		}
-
-		function createContributionArray(type) {
-			const conts = []
-			const d = type.dimensions
-			const baseStep = d * d
-			var k, i = 0
-			while (i < type.pD.length) {
-				const baseSet = type.base[type.pD[i]]
-				let previous, current
-				k = 0
-				do {
-					current = createContribution(type, baseSet, k)
-					if (!previous) {
-						conts[i / baseStep] = current;
-					}
-					else {
-						previous.next = current;
-					}
-					previous = current
-					k += d + 1
-				} while(k < baseSet.length)
-
-				current.next = createContribution(type, type.pD, i + 1)
-				if (d >= 3) {
-					current.next.next = createContribution(type, type.pD, i + d + 2)
-				}
-				if (d === 4) {
-					current.next.next.next = createContribution(type, type.pD, i + 11)
-				}
-				i += baseStep
-			}
-			const result = [conts, createLookupPairs(type.lookup, conts)]
-			type.base = undefined
-			type.lookup = undefined
-			return result
-		}
-
-		let temp = createContributionArray(types._3D)
-		const lookup3D = temp[1]
-		const perm = new Uint8Array(256)
-		const perm3D = new Uint8Array(256)
-		const source = new Uint8Array(setOf(256))
-		var seed = shuffleSeed(clientSeed, 3)
-		doFor(256, function(i) {
-			i = 255 - i
-			seed = shuffleSeed(seed, 1)
-			var r = (seed + 31) % (i + 1)
-			r += r < 0 ? i + 1 : 0
-			perm[i] = source[r]
-			perm3D[i] = perm[i] % 24 * 3
-			source[r] = source[i]
-		})
-		base3D = undefined
-		lookupPairs3D = undefined
-		p3D = undefined
-
-		return function(x, y, z) {
-			const pD = perm3D
-			const p = perm
-			const g = gradients3D
-			const stretchOffset = (x + y + z) * STRETCH_3D
-			const xs = x + stretchOffset, ys = y + stretchOffset, zs = z + stretchOffset
-			const xsb = floor(xs), ysb = floor(ys), zsb = floor(zs)
-			const squishOffset	= (xsb + ysb + zsb) * SQUISH_3D
-			const dx0 = x - (xsb + squishOffset), dy0 = y - (ysb + squishOffset), dz0 = z - (zsb + squishOffset)
-			const xins = xs - xsb, yins = ys - ysb, zins = zs - zsb
-			const inSum = xins + yins + zins
-			var c = lookup3D.get(
-				yins - zins + 1
-				| xins - yins + 1 << 1
-				| xins - zins + 1 << 2
-				| inSum << 3
-				| inSum + zins << 5
-				| inSum + yins << 7
-				| inSum + xins << 9
-			)
-			var i, value = 0
-			while (c !== undefined) {
-				const dx = dx0 + c.dx, dy = dy0 + c.dy, dz = dz0 + c.dz
-				let attn = 2 - dx * dx - dy * dy - dz * dz
-				if (attn > 0) {
-					i = pD[(p[xsb + c.xsb & 0xFF] + (ysb + c.ysb) & 0xFF) + (zsb + c.zsb) & 0xFF]
-					attn *= attn
-					value += attn * attn * (g[i++] * dx + g[i++] * dy + g[i] * dz)
-				}
-				c = c.next
-			}
-			return value * NORM_3D + 0.5
-		}
-	}
-
-	class PVector {
-		constructor(x, y, z) {
-			this.x = x
-			this.y = y
-			this.z = z
-		}
-		set(x, y, z) {
-			if (y === undefined) {
-				this.x = x.x
-				this.y = x.y
-				this.z = x.z
-			}
-			else {
-				this.x = x
-				this.y = y
-				this.z = z
-			}
-		}
-		normalize() {
-			let mag = Math.sqrt(this.x * this.x + this.y * this.y + this.z * this.z)
-			this.x /= mag
-			this.y /= mag
-			this.z /= mag
-		}
-		add(v) {
-			this.x += v.x
-			this.y += v.y
-			this.z += v.z
-		}
-		mult(m) {
-			this.x *= m
-			this.y *= m
-			this.z *= m
-		}
-	}
 	let fill = function(r, g, b) {
 		if (g === undefined) {
 			g = r
@@ -1862,180 +956,9 @@ async function MineKhan() {
 	let sideEdgeBuffers
 	let indexBuffer
 
-	function cross(v1, v2, result) {
-		let x = v1.x,
-			y = v1.y,
-			z = v1.z,
-			x2 = v2.x,
-			y2 = v2.y,
-			z2 = v2.z
-		result.x = y * z2 - y2 * z
-		result.y = z * x2 - z2 * x
-		result.z = x * y2 - x2 * y
-	}
-
 	let matrix = new Float32Array(16); // A temperary matrix that may store random data.
 	let projection = new Float32Array(16)
 	let defaultModelView = new Float32Array([-10,0,0,0,0,10,0,0,0,0,-10,0,0,0,0,1])
-	class Matrix {
-		constructor(arr) {
-			this.elements = new Float32Array(arr || 16)
-		}
-		translate(x, y, z) {
-			let a = this.elements
-			a[3] += a[0] * x + a[1] * y + a[2] * z
-			a[7] += a[4] * x + a[5] * y + a[6] * z
-			a[11] += a[8] * x + a[9] * y + a[10] * z
-			a[15] += a[12] * x + a[13] * y + a[14] * z
-		}
-		rotX(angle) {
-			let elems = this.elements
-			let c = cos(angle)
-			let s = sin(angle)
-			let t = elems[1]
-			elems[1] = t * c + elems[2] * s
-			elems[2] = t * -s + elems[2] * c
-			t = elems[5]
-			elems[5] = t * c + elems[6] * s
-			elems[6] = t * -s + elems[6] * c
-			t = elems[9]
-			elems[9] = t * c + elems[10] * s
-			elems[10] = t * -s + elems[10] * c
-			t = elems[13]
-			elems[13] = t * c + elems[14] * s
-			elems[14] = t * -s + elems[14] * c
-		}
-		rotY(angle) {
-			let c = cos(angle)
-			let s = sin(angle)
-			let elems = this.elements
-			let t = elems[0]
-			elems[0] = t * c + elems[2] * -s
-			elems[2] = t * s + elems[2] * c
-			t = elems[4]
-			elems[4] = t * c + elems[6] * -s
-			elems[6] = t * s + elems[6] * c
-			t = elems[8]
-			elems[8] = t * c + elems[10] * -s
-			elems[10] = t * s + elems[10] * c
-			t = elems[12]
-			elems[12] = t * c + elems[14] * -s
-			elems[14] = t * s + elems[14] * c
-		}
-		scale(x, y, z) {
-			let a = this.elements
-			a[0] *= x;
-			a[1] *= y;
-			a[2] *= z;
-			a[4] *= x;
-			a[5] *= y;
-			a[6] *= z;
-			a[8] *= x;
-			a[9] *= y;
-			a[10] *= z;
-			a[12] *= x;
-			a[13] *= y;
-			a[14] *= z;
-		}
-		identity() {
-			let a = this.elements
-			a[0] = 1
-			a[1] = 0
-			a[2] = 0
-			a[3] = 0
-			a[4] = 0
-			a[5] = 1
-			a[6] = 0
-			a[7] = 0
-			a[8] = 0
-			a[9] = 0
-			a[10] = 1
-			a[11] = 0
-			a[12] = 0
-			a[13] = 0
-			a[14] = 0
-			a[15] = 1
-		}
-		// somebody optimize this
-		// you just have to expand it
-		mult(b) {
-			const a = this.elements.slice()
-			const out = this.elements
-			let e = 0
-			for (let row = 0; row < 4; row++) {
-				for (let col = 0; col < 4; col++) {
-					out[e++] = a[row * 4 + 0] * b[col + 0] + a[row * 4 + 1] * b[col + 4] + a[row * 4 + 2] * b[col + 8] + a[row * 4 + 3] * b[col + 12];
-				}
-			}
-		}
-		// same here
-		postMult(a) {
-			const b = this.elements.slice()
-			const out = this.elements
-			let e = 0
-			for (let row = 0; row < 4; row++) {
-				for (let col = 0; col < 4; col++) {
-					out[e++] = a[row * 4 + 0] * b[col + 0] + a[row * 4 + 1] * b[col + 4] + a[row * 4 + 2] * b[col + 8] + a[row * 4 + 3] * b[col + 12];
-				}
-			}
-		}
-		transpose() {
-			let matrix = this.elements
-			let temp = matrix[4]
-			matrix[4] = matrix[1]
-			matrix[1] = temp
-
-			temp = matrix[8]
-			matrix[8] = matrix[2]
-			matrix[2] = temp
-
-			temp = matrix[6]
-			matrix[6] = matrix[9]
-			matrix[9] = temp
-
-			temp = matrix[3]
-			matrix[3] = matrix[12]
-			matrix[12] = temp
-
-			temp = matrix[7]
-			matrix[7] = matrix[13]
-			matrix[13] = temp
-
-			temp = matrix[11]
-			matrix[11] = matrix[14]
-			matrix[14] = temp
-		}
-		copyArray(from) {
-			let to = this.elements
-			for (let i = 0; i < from.length; i++) {
-				to[i] = from[i]
-			}
-		}
-		copyMatrix(from) {
-			let to = this.elements
-			from = from.elements
-			for (let i = 0; i < from.length; i++) {
-				to[i] = from[i]
-			}
-		}
-	}
-
-	class Plane {
-		constructor(nx, ny, nz) {
-			this.set(nx, ny, nz)
-		}
-		set(nx, ny, nz) {
-			// Pre-computed chunk offsets to reduce branching during culling
-			this.dx = nx > 0 ? 16 : 0
-			this.dy = ny > 0
-			this.dz = nz > 0 ? 16 : 0
-
-			// Normal vector for the plane
-			this.nx = nx
-			this.ny = ny
-			this.nz = nz
-		}
-	}
 
 	let defaultTransformation = new Matrix([-10,0,0,0,0,10,0,0,0,0,-10,0,0,0,0,1])
 	class Camera {
@@ -2192,79 +1115,8 @@ async function MineKhan() {
 		}
 	}
 
-	function trans(matrix, x, y, z) {
-		let a = matrix
-		a[3] += a[0] * x + a[1] * y + a[2] * z
-		a[7] += a[4] * x + a[5] * y + a[6] * z
-		a[11] += a[8] * x + a[9] * y + a[10] * z
-		a[15] += a[12] * x + a[13] * y + a[14] * z
-	}
-	function rotX(matrix, angle) {
-		// This function is basically multiplying 2 4x4 matrices together,
-		// but 1 of them has a bunch of 0's and 1's in it,
-		// so I removed all terms that multiplied by 0, and just left off the 1's.
-		// mat2 = [1, 0, 0, 0, 0, c, -s, 0, 0, s, c, 0, 0, 0, 0, 1]
-		let elems = matrix
-		let c = cos(angle)
-		let s = sin(angle)
-		let t = elems[1]
-		elems[1] = t * c + elems[2] * s
-		elems[2] = t * -s + elems[2] * c
-		t = elems[5]
-		elems[5] = t * c + elems[6] * s
-		elems[6] = t * -s + elems[6] * c
-		t = elems[9]
-		elems[9] = t * c + elems[10] * s
-		elems[10] = t * -s + elems[10] * c
-		t = elems[13]
-		elems[13] = t * c + elems[14] * s
-		elems[14] = t * -s + elems[14] * c
-	}
-	function rotY(matrix, angle) {
-	//source = c, 0, s, 0, 0, 1, 0, 0, -s, 0, c, 0, 0, 0, 0, 1
-		let c = cos(angle)
-		let s = sin(angle)
-		let elems = matrix
-		let t = elems[0]
-		elems[0] = t * c + elems[2] * -s
-		elems[2] = t * s + elems[2] * c
-		t = elems[4]
-		elems[4] = t * c + elems[6] * -s
-		elems[6] = t * s + elems[6] * c
-		t = elems[8]
-		elems[8] = t * c + elems[10] * -s
-		elems[10] = t * s + elems[10] * c
-		t = elems[12]
-		elems[12] = t * c + elems[14] * -s
-		elems[14] = t * s + elems[14] * c
-	}
-	function transpose(matrix) {
-		let temp = matrix[4]
-		matrix[4] = matrix[1]
-		matrix[1] = temp
-
-		temp = matrix[8]
-		matrix[8] = matrix[2]
-		matrix[2] = temp
-
-		temp = matrix[6]
-		matrix[6] = matrix[9]
-		matrix[9] = temp
-
-		temp = matrix[3]
-		matrix[3] = matrix[12]
-		matrix[12] = temp
-
-		temp = matrix[7]
-		matrix[7] = matrix[13]
-		matrix[13] = temp
-
-		temp = matrix[11]
-		matrix[11] = matrix[14]
-		matrix[14] = temp
-	}
 	function matMult() {
-	//Multiply the projection matrix by the view matrix; this is optimized specifically for these matrices by removing terms that are always 0.
+		//Multiply the projection matrix by the view matrix; this is optimized specifically for these matrices by removing terms that are always 0.
 		let proj = projection
 		let view = modelView
 		matrix[0] = proj[0] * view[0]
@@ -2284,11 +1136,7 @@ async function MineKhan() {
 		matrix[14] = proj[14] * view[10]
 		matrix[15] = proj[14] * view[11]
 	}
-	function copyArr(a, b) {
-		for (let i = 0; i < a.length; i++) {
-			b[i] = a[i]
-		}
-	}
+
 	function FOV(fov) {
 		let tang = Math.tan(fov * 0.5 * Math.PI / 180)
 		let scale = 1 / tang
@@ -2301,6 +1149,7 @@ async function MineKhan() {
 		projection[11] = -1
 		projection[14] = -far * near / (far - near)
 	}
+
 	function initModelView(camera, x, y, z, rx, ry) {
 		if (camera) {
 			camera.transform()
@@ -2317,45 +1166,6 @@ async function MineKhan() {
 		}
 	}
 
-	function timeString(millis) {
-		if (millis > 300000000000 || !millis) {
-			return "never"
-		}
-		const SECOND = 1000
-		const MINUTE = SECOND * 60
-		const HOUR = MINUTE * 60
-		const DAY = HOUR * 24
-		const YEAR = DAY * 365
-
-		if (millis < MINUTE) {
-			return "just now"
-		}
-
-		let years = floor(millis / YEAR)
-		millis -= years * YEAR
-
-		let days = floor(millis / DAY)
-		millis -= days * DAY
-
-		let hours = floor(millis / HOUR)
-		millis -= hours * HOUR
-
-		let minutes = floor(millis / MINUTE)
-
-		if (years) {
-			return `${years} year${years > 1 ? "s" : ""} and ${days} day${days !== 1 ? "s" : ""} ago`
-		}
-		if (days) {
-			return `${days} day${days > 1 ? "s" : ""} and ${hours} hour${hours !== 1 ? "s" : ""} ago`
-		}
-		if (hours) {
-			return `${hours} hour${hours > 1 ? "s" : ""} and ${minutes} minute${minutes !== 1 ? "s" : ""} ago`
-		}
-		return `${minutes} minute${minutes > 1 ? "s" : ""} ago`
-	}
-	function roundBits(number) {
-		return (number * 1000000 + 0.5 | 0) / 1000000
-	}
 	function rayTrace(x, y, z, shape) {
 		let cf, cd = 1e9; //Closest face and distance
 		let m; //Absolute distance to intersection point
@@ -5767,7 +4577,7 @@ async function MineKhan() {
 		textureMap = {}
 		textureCoords = []
 
-		setPixel = function(textureNum, x, y, r, g, b, a) {
+		const setPixel = function(textureNum, x, y, r, g, b, a) {
 			let texX = textureNum & 15
 			let texY = textureNum >> 4
 			let offset = (texY * 16 + y) * 1024 + texX * 64 + x * 4
@@ -5776,7 +4586,7 @@ async function MineKhan() {
 			texturePixels[offset + 2] = b
 			texturePixels[offset + 3] = a !== undefined ? a : 255
 		}
-		getPixels = function(str) {
+		const getPixels = function(str) {
 			// var w = parseInt(str.substr(0, 2), 36)
 			// var h = parseInt(str.substr(2, 2), 36)
 			var colors = []
@@ -5794,6 +4604,8 @@ async function MineKhan() {
 			}
 			return pixels
 		};
+
+		const textures = texturesFunc(setPixel, getPixels);
 
 		{
 			// Specify the texture coords for each index
